Изменить Map<UUID, Integer> products на Map<UUID, Long>

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/BookedProductsDto.java
@@ -19,7 +19,7 @@ public class BookedProductsDto {
 
     boolean fragile;
 
-    Map<UUID, Integer> products;
+    Map<UUID, Long> products;
 
     UUID orderId;
